# openapi.yaml
openapi: 3.1.0
info:
  title: Agentic Voice SDR API
  version: 1.0.0
  description: |
    Esta API permite configurar, testar e operar um agente de vendas por voz baseado em inteligência artificial. O agente interpreta o que o lead disse, decide o próximo passo comercial e gera uma resposta falada. A API permite ajustar e testar o comportamento antes de realizar chamadas reais.

    Com esta API você pode configurar o comportamento e as instruções do agente, escolher e ajustar a voz, configurar decisões comerciais, testar frases sem telefonar, ouvir o áudio gerado, medir o tempo de resposta, consultar o estado dos serviços, iniciar e acompanhar chamadas e consultar transcrições.

    Fluxo recomendado: (1) faça login; (2) consulte GET /v1/runtime/status para verificar o estado do sistema; (3) confira ou altere o prompt; (4) ajuste decisão comercial, voz e estilo de fala; (5) teste uma frase de lead; (6) confira a decisão e execute um teste de resposta do agente; (7) ouça o áudio; (8) ajuste configurações se necessário; (9) somente depois realize uma chamada real. Os testes sem chamada permitem avaliar intenção, próximo passo, latência, resposta gerada e áudio.

    O mecanismo de decisão comercial seleciona um próximo passo, como continuar, perguntar, tratar uma objeção, propor uma ação, fazer follow-up, encerrar ou solicitar atendimento humano. Em seguida, o modelo de voz gera a resposta. Isso separa estratégia comercial de personalidade, estilo, voz, ritmo e tom.

...

...

...
servers:
...
paths:
...
  /v1/config/whatsapp:
    put:
      operationId: configureWhatsApp
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WhatsAppConfigRequest' }
      responses:
        '200': { description: Safe configuration, content: { application/json: { schema: { $ref: '#/components/schemas/WhatsAppConfig' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '502': { $ref: '#/components/responses/ProviderFailure' }
    get:
      operationId: getWhatsApp
      responses:
        '200': { description: Safe configuration, content: { application/json: { schema: { $ref: '#/components/schemas/WhatsAppConfig' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
  /v1/config/whatsapp/instances:
    get:
      operationId: listWhatsAppInstances
      responses:
        '200': { description: Provider instances, content: { application/json: { schema: { type: object, required: [instances], properties: { instances: { type: array, items: { $ref: '#/components/schemas/WhatsAppInstance' } } } } } } }
        '502': { $ref: '#/components/responses/ProviderFailure' }
  /v1/config/whatsapp/instance:
    put:
      operationId: selectWhatsAppInstance
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [instance_id], properties: { instance_id: { type: string, minLength: 1 } } }
      responses:
        '200': { description: Selected safe configuration, content: { application/json: { schema: { $ref: '#/components/schemas/WhatsAppConfig' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
  /v1/config/whatsapp/test:
    post:
      operationId: testWhatsApp
      description: Validates the configured provider and active instance; sends no message. Body is intentionally empty.
      responses:
        '200': { description: Test result }
        '400': { $ref: '#/components/responses/BadRequest' }
        '502': { $ref: '#/components/responses/ProviderFailure' }
...
components:
  securitySchemes:
    BearerAuth: { type: http, scheme: bearer }
  responses:
    BadRequest: { description: Invalid request, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
    ProviderFailure: { description: Safe provider failure, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  schemas:
    Error: { type: object, required: [error], properties: { error: { type: string, description: Safe summary; raw provider messages are omitted. }, provider_status_class: { type: string, description: Optional sanitized provider operation/transport category. } } }
...
    WhatsAppConfigRequest:
      type: object
      required: [provider, base_url, credential]
      properties:
        provider: { type: string }
        base_url: { type: string, format: uri }
        credential: { type: string, writeOnly: true }
    WhatsAppConfig:
      type: object
      required: [provider, base_url, credential_configured, sdr_status]
      properties:
        provider: { type: string }
        base_url: { type: string, format: uri }
        credential_configured: { type: boolean }
        active_instance_id: { type: string }
        active_instance_phone: { type: string }
        provider_status: { type: string }
        sdr_status: { type: string }
        last_verified_at: { type: string, format: date-time }
    WhatsAppInstance:
      type: object
      required: [id, status]
      properties:
        id: { type: string }
        name: { type: string }
        phone: { type: string }
        status: { type: string, enum: [CONNECTED, READY, DISCONNECTED] }
...